requirements.txt
#Python Dependencies for the Django Backend

Django==5.0.0
djangorestframework==3.15.0
psycopg==3.2.13
django-cors-headers==4.3.1
pandas